IBX-10636: Adjust twig components to new standards
