+++ src/bundle/Resources/views/themes/standard/design_system/components/accordion.html.twig
@@ -15,7 +15,7 @@
 
             {% endblock header %}
         </div>
-        <twig:ibexa:Expander
+        <twig:ibexa:expander
             type="caret"
             :hasIcon="true"
             :isExpanded="initially_expanded"
