+++ src/bundle/Resources/views/themes/standard/design_system/components/form_controls/input_text.html.twig
@@ -1,5 +1,5 @@
 {% extends '@IbexaDesignSystemTwig/themes/standard/design_system/partials/base_form_control.html.twig' %}
 
 {% block content %}
-    <twig:ibexa:inputs:InputText {{ ...input }} />
+    <twig:ibexa:inputs:input_text {{ ...input }} />
 {% endblock content %}
