+++ src/bundle/Resources/views/themes/standard/design_system/components/helper_text.html.twig
@@ -12,12 +12,12 @@
 
 <div
     class="{{
-    helper_text_classes.apply(
-        {
-            type
-        },
-        attributes.render('class')
-    )
+        helper_text_classes.apply(
+            {
+                type
+            },
+            attributes.render('class')
+        )
     }}"
     {{ attributes }}
 >
