+++ src/bundle/Resources/views/themes/standard/design_system/components/icon.html.twig
@@ -19,12 +19,12 @@
 
 <svg
     class="{{
-    icon_classes.apply(
-        {
-            size
-        },
-        attributes.render('class')
-    )
+        icon_classes.apply(
+            {
+                size
+            },
+            attributes.render('class')
+        )
     }}"
     {{ attributes }}
 >
