+++ src/bundle/Resources/views/themes/standard/design_system/components/icon_button.html.twig
@@ -1,4 +1,4 @@
-{% extends '@IbexaDesignSystemTwig/themes/standard/design_system/components/Button.html.twig' %}
+{% extends '@IbexaDesignSystemTwig/themes/standard/design_system/components/button.html.twig' %}
 
 {% block button_inner %}
     <div class="ids-btn__icon">
