+++ src/bundle/Resources/views/themes/standard/design_system/components/inputs/radio_button.html.twig
@@ -15,10 +15,10 @@
         class="{{ input_classes }}"
         type="radio"
         {{
-        attributes.defaults({
-            disabled,
-            required
-        })
+            attributes.defaults({
+                disabled,
+                required
+            })
         }}
     />
 </div>
