+++ src/bundle/Resources/views/themes/standard/design_system/components/inputs/three_state_checkbox.html.twig
@@ -4,21 +4,22 @@
         {
             'ids-input--disabled': disabled,
             'ids-input--error': error,
-            'ids-input--required': required
+            'ids-input--required': required,
+            'ids-input--indeterminate': indeterminate
         },
         attributes.render('class') ?? ''
     )
 %}
 
-<div class="ids-checkbox">
+<div class="ids-three-state-checkbox">
     <input
         class="{{ input_classes }}"
         type="checkbox"
         {{
-        attributes.defaults({
-            disabled,
-            required
-        })
+            attributes.defaults({
+                disabled,
+                required
+            })
         }}
     />
 </div>
