+++ src/lib/Twig/Components/Inputs/RadioButton.php
@@ -6,12 +6,12 @@
  */
 declare(strict_types=1);
 
-namespace Ibexa\DesignSystemTwig\Twig\Components\inputs;
+namespace Ibexa\DesignSystemTwig\Twig\Components\Inputs;
 
 use Symfony\Component\OptionsResolver\OptionsResolver;
 use Symfony\UX\TwigComponent\Attribute\AsTwigComponent;
 
-#[AsTwigComponent]
+#[AsTwigComponent('ibexa:inputs:radio_button')]
 final class RadioButton extends AbstractChoiceInput
 {
     protected function configurePropsResolver(OptionsResolver $resolver): void
