+++ src/lib/Twig/Components/Inputs/ThreeStateCheckbox.php
@@ -6,15 +6,21 @@
  */
 declare(strict_types=1);
 
-namespace Ibexa\DesignSystemTwig\Twig\Components\inputs;
+namespace Ibexa\DesignSystemTwig\Twig\Components\Inputs;
 
 use Symfony\Component\OptionsResolver\OptionsResolver;
 use Symfony\UX\TwigComponent\Attribute\AsTwigComponent;
 
-#[AsTwigComponent]
-final class Checkbox extends AbstractChoiceInput
+#[AsTwigComponent('ibexa:inputs:three_state_checkbox')]
+final class ThreeStateCheckbox extends AbstractChoiceInput
 {
+    public bool $indeterminate = false;
+
     protected function configurePropsResolver(OptionsResolver $resolver): void
     {
+        $resolver
+            ->define('indeterminate')
+            ->allowedTypes('bool')
+            ->default(false);
     }
 }
